refactor(ui): padroniza cores com variáveis e extrai componente FormField global
